seperated the shelf from the index file and imported it to keep things cleaner
